moved dapi endpoint to helpers
